requirements.txt
grpcio
grpcio-tools
transformers
datasets
numpy
torch